requirements.txt
opencv-python
numpy
imageio
Pillow